tsBlockCipher: export the Instance() static method
PROPERTIES() is used by some templates to create a class.
It needs to be exported so it can be accessed through DLLs.

With some compiler optimization PROPERTIES() becomes inline,
even with a dllimport in front. In that case the Instance()
also needs to exported so it can be accessed through DLLs.